update e horizon mpp functions so they return a list of list of edges
