+++ libnavigation-base/api/current.txt
@@ -190,125 +190,6 @@ package com.mapbox.navigation.base.time.span {
 
 package com.mapbox.navigation.base.trip.model {
 
-  public final class EHorizon {
-    method public com.mapbox.navigation.base.trip.model.Edge current(com.mapbox.navigation.base.trip.model.EHorizonPosition position);
-    method public com.mapbox.navigation.base.trip.model.Edge getStart();
-    method public java.util.List<com.mapbox.navigation.base.trip.model.Edge> mpp(com.mapbox.navigation.base.trip.model.EHorizonPosition position);
-    method public java.util.List<com.mapbox.navigation.base.trip.model.Edge> mpp();
-    method public com.mapbox.navigation.base.trip.model.EHorizon.Builder toBuilder();
-  }
-
-  public static final class EHorizon.Builder {
-    ctor public EHorizon.Builder();
-    method public com.mapbox.navigation.base.trip.model.EHorizon build();
-    method public com.mapbox.navigation.base.trip.model.EHorizon.Builder start(com.mapbox.navigation.base.trip.model.Edge start);
-  }
-
-  public final class EHorizonPosition {
-    method public long getEdgeId();
-    method public double getPercentAlong();
-    method public com.mapbox.navigation.base.trip.model.EHorizonPosition.Builder toBuilder();
-  }
-
-  public static final class EHorizonPosition.Builder {
-    ctor public EHorizonPosition.Builder();
-    method public com.mapbox.navigation.base.trip.model.EHorizonPosition build();
-    method public com.mapbox.navigation.base.trip.model.EHorizonPosition.Builder edgeId(long edgeId);
-    method public com.mapbox.navigation.base.trip.model.EHorizonPosition.Builder percentAlong(double percentAlong);
-  }
-
-  public final class EHorizonResultType {
-    field public static final String INITIAL = "INITIAL";
-    field public static final com.mapbox.navigation.base.trip.model.EHorizonResultType! INSTANCE;
-    field public static final String UPDATE = "UPDATE";
-  }
-
-  public final class Edge {
-    method public <R> R! accept(com.mapbox.navigation.base.trip.model.EdgeVisitor<? extends R> visitor);
-    method public <R> java.util.List<R> collect(kotlin.jvm.functions.Function1<? super com.mapbox.navigation.base.trip.model.Edge,? extends R> fn);
-    method public boolean getBridge();
-    method public String? getCountryCode();
-    method public byte getCurvature();
-    method public String getFrc();
-    method public com.mapbox.geojson.LineString? getGeometry();
-    method public double getHeading();
-    method public long getId();
-    method public Byte? getLaneCount();
-    method public double getLength();
-    method public byte getLevel();
-    method public Double? getMeanElevation();
-    method public boolean getMotorway();
-    method public java.util.List<com.mapbox.navigation.base.trip.model.NameInfo> getNames();
-    method public java.util.List<com.mapbox.navigation.base.trip.model.Edge> getOut();
-    method public boolean getPositiveDirection();
-    method public double getProbability();
-    method public boolean getRamp();
-    method public double getSpeed();
-    method public Double? getSpeedLimit();
-    method public String? getStateCode();
-    method public boolean getToll();
-    method public boolean getTunnel();
-    method public String getWayId();
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder toBuilder();
-  }
-
-  public static final class Edge.Builder {
-    ctor public Edge.Builder();
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder bridge(boolean bridge);
-    method public com.mapbox.navigation.base.trip.model.Edge build();
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder countryCode(String? countryCode);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder curvature(byte curvature);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder frc(String frc);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder geometry(com.mapbox.geojson.LineString? geometry);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder heading(double heading);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder id(long id);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder laneCount(Byte? laneCount);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder length(double length);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder level(byte level);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder meanElevation(Double? meanElevation);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder motorway(boolean motorway);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder names(java.util.List<com.mapbox.navigation.base.trip.model.NameInfo> names);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder out(java.util.List<com.mapbox.navigation.base.trip.model.Edge> out);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder positiveDirection(boolean positiveDirection);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder probability(double probability);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder ramp(boolean ramp);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder speed(double speed);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder speedLimit(Double? speedLimit);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder stateCode(String? stateCode);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder toll(boolean toll);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder tunnel(boolean tunnel);
-    method public com.mapbox.navigation.base.trip.model.Edge.Builder wayId(String wayId);
-  }
-
-  public interface EdgeVisitor<R> {
-    method public R! visit(com.mapbox.navigation.base.trip.model.Edge edge);
-  }
-
-  public final class FunctionalRoadClass {
-    field public static final com.mapbox.navigation.base.trip.model.FunctionalRoadClass! INSTANCE;
-    field public static final String MOTORWAY = "MOTORWAY";
-    field public static final String PRIMARY = "PRIMARY";
-    field public static final String RESIDENTIAL = "RESIDENTIAL";
-    field public static final String SECONDARY = "SECONDARY";
-    field public static final String SERVICE_OTHER = "SERVICE_OTHER";
-    field public static final String TERTIARY = "TERTIARY";
-    field public static final String TRUNK = "TRUNK";
-    field public static final String UNCLASSIFIED = "UNCLASSIFIED";
-  }
-
-  public final class NameInfo {
-    method public String getName();
-    method public boolean getShielded();
-    method public com.mapbox.navigation.base.trip.model.NameInfo.Builder toBuilder();
-  }
-
-  public static final class NameInfo.Builder {
-    ctor public NameInfo.Builder();
-    method public com.mapbox.navigation.base.trip.model.NameInfo build();
-    method public com.mapbox.navigation.base.trip.model.NameInfo.Builder name(String name);
-    method public com.mapbox.navigation.base.trip.model.NameInfo.Builder shielded(boolean shielded);
-  }
-
   public final class RouteLegProgress {
     method public com.mapbox.navigation.base.trip.model.RouteStepProgress? getCurrentStepProgress();
     method public float getDistanceRemaining();
