+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/MapboxNavigation.kt
@@ -613,11 +613,6 @@ class MapboxNavigation(
      *
      * Registering an EHorizonObserver activates the Electronic Horizon module.
      *
-     * Electronic Horizon is still **experimental**, which means that the design of the
-     * APIs has open issues which may (or may not) lead to their changes in the future.
-     * Roughly speaking, there is a chance that those declarations will be deprecated in the near
-     * future or the semantics of their behavior may change in some way that may break some code.
-     *
      * @see unregisterEHorizonObserver
      */
     fun registerEHorizonObserver(eHorizonObserver: EHorizonObserver) {
@@ -628,11 +623,6 @@ class MapboxNavigation(
      * Unregisters a EHorizon observer.
      *
      * Unregistering all observers deactivates the module.
-     *
-     * Electronic Horizon is still **experimental**, which means that the design of the
-     * APIs has open issues which may (or may not) lead to their changes in the future.
-     * Roughly speaking, there is a chance that those declarations will be deprecated in the near
-     * future or the semantics of their behavior may change in some way that may break some code.
      *
      * @see registerEHorizonObserver
      */
