+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/model/eh/FunctionalRoadClass.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.base.trip.model
+package com.mapbox.navigation.core.trip.model.eh
 
 /**
  * Holds available [FunctionalRoadClass] types. See https://wiki.openstreetmap.org/wiki/Key:highway for further details.
@@ -12,11 +12,6 @@ package com.mapbox.navigation.base.trip.model
  * - [FunctionalRoadClass.UNCLASSIFIED]
  * - [FunctionalRoadClass.RESIDENTIAL]
  * - [FunctionalRoadClass.SERVICE_OTHER]
- *
- * Electronic Horizon is still **experimental**, which means that the design of the
- * APIs has open issues which may (or may not) lead to their changes in the future.
- * Roughly speaking, there is a chance that those declarations will be deprecated in the near
- * future or the semantics of their behavior may change in some way that may break some code.
  */
 object FunctionalRoadClass {
