+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/model/eh/NameInfo.kt
@@ -1,0 +1,47 @@
+package com.mapbox.navigation.core.trip.model.eh
+
+/**
+ * Road name information
+ *
+ * @param name road name
+ * @param shielded is the road shielded?
+ */
+class NameInfo internal constructor(
+    val name: String,
+    val shielded: Boolean
+) {
+
+    /**
+     * Regenerate whenever a change is made
+     */
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as NameInfo
+
+        if (name != other.name) return false
+        if (shielded != other.shielded) return false
+
+        return true
+    }
+
+    /**
+     * Regenerate whenever a change is made
+     */
+    override fun hashCode(): Int {
+        var result = name.hashCode()
+        result = 31 * result + shielded.hashCode()
+        return result
+    }
+
+    /**
+     * Returns a string representation of the object.
+     */
+    override fun toString(): String {
+        return "NameInfo(" +
+            "name=$name, " +
+            "shielded=$shielded" +
+            ")"
+    }
+}
